requirements.txt
python-dotenv
PyPDF2
streamlit
langchain
openai
faiss-cpu

